# config/openapi_specs/政府開放資料測試平台_openapi.json
{"openapi":"3.1.0","info":{"title":"政府開放資料測試平台 API","description":"用於與 GovOpenApiAgent 進行對測的政府開放資料測試平台 API。","version":"1.0.0"},"servers":[{"url":"http://localhost:8000","description":"Local Development Server"}],"paths":{"/opendata":{"get":{"summary":"取得所有開放資料列表","operationId":"get_all_opendata_opendata_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"sort_by","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"排序欄位 (例如：name, department, update_date)","title":"Sort By"},"description":"排序欄位 (例如：name, department, update_date)"},{"name":"order","in":"query","required":false,"schema":{"type":"string","pattern":"^(asc|desc)$","description":"排序順序 (asc 或 desc)","default":"asc","title":"Order"},"description":"排序順序 (asc 或 desc)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenDataListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/opendata/search":{"get":{"summary":"搜尋開放資料","operationId":"search_opendata_opendata_search_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","minLength":1,"description":"搜尋關鍵字，可匹配資料集名稱、資料集描述、提供機關等","title":"Query"},"description":"搜尋關鍵字，可匹配資料集名稱、資料集描述、提供機關等"},{"name":"department","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"依提供機關名稱篩選","title":"Department"},"description":"依提供關名稱篩選"},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"依服務分類篩選","title":"Category"},"description":"依服務分類篩選"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"sort_by","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"排序欄位 (例如：name, department, update_date)","title":"Sort By"},"description":"排序欄位 (例如：name, department, update_date)"},{"name":"order","in":"query","required":false,"schema":{"type":"string","pattern":"^(asc|desc)$","description":"排序順序 (asc 或 desc)","default":"asc","title":"Order"},"description":"排序順序 (asc 或 desc)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenDataListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/opendata/categories":{"get":{"summary":"取得開放資料類別列表","operationId":"get_opendata_categories_opendata_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"string"},"type":"array","title":"Response Get Opendata Categories Opendata Categories Get"}}}}}}},"/opendata/departments":{"get":{"summary":"取得開放資料提供單位列表","operationId":"get_opendata_departments_opendata_departments_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"string"},"type":"array","title":"Response Get Opendata Departments Opendata Departments Get"}}}}}}},"/opendata/{item_id}":{"get":{"summary":"取得單一開放資料詳情","operationId":"get_opendata_item_opendata__item_id__get","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenDataItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OpenDataItem":{"properties":{"資料集識別碼":{"type":"string","title":"資料集識別碼"},"資料集名稱":{"type":"string","title":"資料集名稱"},"提供機關":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"提供機關"},"資料集描述":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"資料集描述"},"資料下載網址":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"資料下載網址"},"服務分類":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"服務分類"},"詮釋資料更新時間":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"詮釋資料更新時間"},"資料提供屬性":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"資料提供屬性"},"品質檢測":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"品質檢測"},"檔案格式":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"檔案格式"},"編碼格式":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"編碼格式"},"資資料集上架方式":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"資資料集上架方式"},"主要欄位說明":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"主要欄位說明"},"更新頻率":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"更新頻率"},"授權方式":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"授權方式"},"相關網址":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"相關網址"},"計費方式":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"計費方式"},"提供機關聯絡人姓名":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"提供機關聯絡人姓名"},"提供機關聯絡人電話":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"提供機關聯絡人電話"},"上架日期":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"上架日期"},"備註":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"備註"},"資料量":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"資料量"}},"type":"object","required":["資料集識別碼","資料集名稱"],"title":"OpenDataItem"},"OpenDataListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/OpenDataItem"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["items","total"],"title":"OpenDataListResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}